# Fix ServiceController name population perf