--- a/src/System.ServiceProcess.ServiceController/src/Resources/Strings.resx
+++ b/src/System.ServiceProcess.ServiceController/src/Resources/Strings.resx
@@ -62,10 +62,10 @@
     <value>Arguments within the 'args' array passed to Start cannot be null.</value>
   </data>
   <data name="BadMachineName" xml:space="preserve">
-    <value>MachineName value {0} is invalid.</value>
+    <value>MachineName '{0}' is invalid.</value>
   </data>
   <data name="CannotStart" xml:space="preserve">
-    <value>Cannot start service {0} on computer '{1}'.</value>
+    <value>Cannot start service '{0}' on computer '{1}'.</value>
   </data>
   <data name="InvalidEnumArgument" xml:space="preserve">
     <value>The value of argument '{0}' ({1}) is invalid for Enum type '{2}'.</value>
@@ -77,22 +77,22 @@
     <value>MachineName was not set.</value>
   </data>
   <data name="NoService" xml:space="preserve">
-    <value>Service {0} was not found on computer '{1}'.</value>
+    <value>Service '{0}' was not found on computer '{1}'.</value>
   </data>
   <data name="OpenSC" xml:space="preserve">
     <value>Cannot open Service Control Manager on computer '{0}'. This operation might require other privileges.</value>
   </data>
   <data name="OpenService" xml:space="preserve">
-    <value>Cannot open {0} service on computer '{1}'.</value>
+    <value>Cannot open '{0}' service on computer '{1}'.</value>
   </data>
   <data name="PauseService" xml:space="preserve">
-    <value>Cannot pause {0} service on computer '{1}'.</value>
+    <value>Cannot pause '{0}' service on computer '{1}'.</value>
   </data>
   <data name="ResumeService" xml:space="preserve">
-    <value>Cannot resume {0} service on computer '{1}'.</value>
+    <value>Cannot resume '{0}' service on computer '{1}'.</value>
   </data>
   <data name="StopService" xml:space="preserve">
-    <value>Cannot stop {0} service on computer '{1}'.</value>
+    <value>Cannot stop '{0}' service on computer '{1}'.</value>
   </data>
   <data name="Timeout" xml:space="preserve">
     <value>Time out has expired and the operation has not been completed.</value>
@@ -107,7 +107,7 @@
     <value>Cannot change service name when the service is running.</value>
   </data>
   <data name="ServiceName" xml:space="preserve">
-    <value>Service name {0} contains invalid characters, is empty, or is too long (max length = {1}).</value>
+    <value>Service name '{0}' contains invalid characters, is empty, or is too long (max length = {1}).</value>
   </data>
   <data name="NoServices" xml:space="preserve">
     <value>Service has not been supplied. At least one object derived from ServiceBase is required in order to run.</value>
@@ -179,18 +179,18 @@
     <value>Failed in handling the PowerEvent. The error that occurred was: {0}.</value>
   </data>
   <data name="InstallOK" xml:space="preserve">
-    <value>Service {0} has been successfully installed.</value>
+    <value>Service '{0}' has been successfully installed.</value>
   </data>
   <data name="TryToStop" xml:space="preserve">
-    <value>Attempt to stop service {0}.</value>
+    <value>Attempt to stop service '{0}'.</value>
   </data>
   <data name="ServiceRemoving" xml:space="preserve">
-    <value>Service {0} is being removed from the system...</value>
+    <value>Service '{0}' is being removed from the system...</value>
   </data>
   <data name="ServiceRemoved" xml:space="preserve">
-    <value>Service {0} was successfully removed from the system.</value>
+    <value>Service '{0}'' was successfully removed from the system.</value>
   </data>
   <data name="ControlService" xml:space="preserve">
-    <value>Cannot control {0} service on computer '{1}'.</value>
+    <value>Cannot control '{0}' service on computer '{1}'.</value>
   </data>
 </root>