Fix https://github.com/dotnet/corefx/issues/31877

* Regression was caused by an unexplained change in Project K in 2014 to populate name and display name by enumerating every service on the machine examining their names. Revert back to use the direct Win32 API calls. Modernize the interop.
* Tidy up some strings.
* Write tests to get code coverage of various name population codepaths.
* Fix several existing issues uncovered by tests, in some cases introduced since .NET Framework.
* Inline one method for clarity.
```
BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17134.228 (1803/April2018Update/Redstone4)
Intel Core i7-2760QM CPU 2.40GHz (Sandy Bridge), 1 CPU, 8 logical and 4 physical cores
Frequency=2336175 Hz, Resolution=428.0501 ns, Timer=TSC
.NET Core SDK=2.1.400
  [Host] : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT
  Clr    : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3132.0
  Core   : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT
```
 Method |  Job | Runtime |     Mean |     Error |    StdDev |
------- |----- |-------- |---------:|----------:|----------:|
   GetStatus |  Clr |     Clr | 195.6 us | 3.8735 us | 9.8594 us |
   GetStatus| Core |    Core | 2,588.3 us | 14.756 us | 11.521 us |
   GetStatusFixed | Core |    Core | 181.6 us | 0.7662 us | 0.6398 us |

I also noticed that .NET Framework was using raw IntPtr for the SCM handle but apparently that doesn't help it significantly:

Method |     Mean |    Error |   StdDev |
------------------ |---------:|---------:|---------:|
IntPtr | 186.8 us | 2.791 us | 2.611 us |
 SafeServiceHandle | 186.2 us | 2.077 us | 1.842 us |