@danmosemsft 2018-09-03
@dotnet-bot test UWP CoreCLR x64 Debug Build